make changes necessary for BotanicalCorrections
@gabrielareto I am pushing your function along with changes necessary in the package.
But, as of now, it is not live in the App because it does not have any arguments other that "Data".
I'll try to find a way around that (the easiest is to add an argument to the function, so let me know if you have one in mind that would be useful).